requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain_google_genai
python-dotenv
typesense
langchain_openai
langgraph
streamlit
